+++ sgs-server/src/test/java/com/sun/sgs/test/impl/service/nodemap/affinity/TestGraphs.java
@@ -223,14 +223,14 @@ public class TestGraphs {
         }
 
         /** {@inheritDoc} */
-        public ConcurrentHashMap<Long, ConcurrentHashMap<Object, Integer>>
+        public ConcurrentHashMap<Long, ConcurrentHashMap<Object, Long>>
                 getConflictMap() 
         {
             throw new UnsupportedOperationException("Not supported yet.");
         }
 
         /** {@inheritDoc} */
-        public Map<Object, Map<Identity, Integer>> getObjectUseMap() {
+        public Map<Object, Map<Identity, Long>> getObjectUseMap() {
             throw new UnsupportedOperationException("Not supported yet.");
         }
